#37 - Moved puzzle data to separate mjs file
